wineQualityWithKNN
This class estimates the wine quality using the k-nearest neighbor (KNN) algorithm and recommends wine to the user from the data set using the KNN algorithm based on the values entered by the user. The class uses the red wine quality data from kaggle. (Kaggle link is below.) Also, there are graphs for various purposes in this class.